// src/domains/Dashboard/DashboardChart/enum/DashboardChart.enum.ts
export enum BarStackId {
    Stack = 'stack',
}

export enum ChartDataKey {
    Investment = 'investment',
    Loss = 'loss',
    Profit = 'profit',
    Maintenance = 'maintenance',
    Name = 'name',
}

export enum TimeFrame {
    Today = 'today',
    ThisMonth = 'thisMonth',
    ThisYear = 'thisYear',
}

export enum BarSize {
    Max = 30,
}